The functionalities from HostedDeviceService (plugins layer) have been moved into the Application layer because starting and stopping an emulator is actually a use case. For achieving this a GenericDevice.cs (application layer) has been created that contains the start, stop and command execution composition logic.
The Program.cs does no longer contain configuration logic. The configuration logic has been moved to the application layer (GenericDevice.cs).
The IDeviceConfiguration.cs has been split to separate configurations for the IServer.cs (IServerConfiguration.cs).

As the TcpServer.cs is a plugin an IServer interface is now present in the application layer and implemented by the TcpServer.cs to allow the application to control the server.
The IByteArrayConvertible.cs interface is now an IStringConvertible.cs interface that does not anything know about byte arrays because that is a socket server specific requirement.

Namespaces have been adjusted.

Adapters layer has been removed as it is not required. The IProtocolInterpreter.cs was moved to the application layer because the GenericDevice.cs has to interact with it (moved from adapters; dependency rule). The implementation of the IProtocolInterpreter.cs is a part of the plugins layer because removing the antlr parser also requires rewriting all visitors and interpreter logic.